Fix null timestamp / formatting issues
