+++ api/src/main/java/com/codeforcommunity/dto/announcements/Announcement.java
@@ -1,5 +1,6 @@
 package com.codeforcommunity.dto.announcements;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import java.sql.Timestamp;
 
 public class Announcement {
@@ -7,6 +8,8 @@ public class Announcement {
   private int id;
   private String title;
   private String description;
+
+  @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss Z")
   private Timestamp created;
 
   private Announcement() {}
@@ -32,5 +35,15 @@ public class Announcement {
 
   public Timestamp getCreated() {
     return created;
+  }
+
+  @Override
+  public String toString() {
+    return "Announcement{" +
+        "id=" + id +
+        ", title='" + title + '\'' +
+        ", description='" + description + '\'' +
+        ", created=" + created +
+        '}';
   }
 }
