+++ api/src/main/java/com/codeforcommunity/dto/announcements/PostAnnouncementResponse.java
@@ -4,44 +4,15 @@ import java.sql.Timestamp;
 
 public class PostAnnouncementResponse {
 
-  private int id;
-  private String title;
-  private String description;
-  private Timestamp created;
+  private Announcement announcement;
 
   private PostAnnouncementResponse() {}
 
-  public PostAnnouncementResponse(int id, String title, String description,
-      Timestamp created) {
-    this.id = id;
-    this.title = title;
-    this.description = description;
-    this.created = created;
+  public PostAnnouncementResponse(Announcement announcement) {
+    this.announcement = announcement;
   }
 
-  public int getId() {
-    return id;
-  }
-
-  public String getTitle() {
-    return title;
-  }
-
-  public String getDescription() {
-    return description;
-  }
-
-  public Timestamp getCreated() {
-    return created;
-  }
-
-  @Override
-  public String toString() {
-    return "PostAnnouncementsResponse{" +
-        "id=" + id +
-        ", title='" + title + '\'' +
-        ", description='" + description + '\'' +
-        ", created=" + created +
-        '}';
+  public Announcement getAnnouncement() {
+    return announcement;
   }
 }
